move file reading into input module
