+++ src/input/mod.rs
@@ -1,1 +1,2 @@
+pub mod polyanya_file;
 pub mod trimesh;
